fix delete range code; kludgy step noted
On testing, closing the iterator after a "delete all" will cause a panic, whereas not
closing the iterator works fine.  Since the # of delete calls will be very limited and
the iterator size is negligable, I am simply not closing the iterator after doing
some work trying to find underlying issues.  It could be a driver issue or an interaction
between deleted k/v and the underlying iterator, which shouldn't be case from leveldb
boards.  Due to lack of dev time, we note this issue and move on.